requirements.txt
python-telegram-bot==21.6
pytesseract==0.3.13
Pillow==10.4.0
